+++ pywren_ibm_cloud/compute/backends/openwhisk/entry_point.py
@@ -18,8 +18,8 @@ import os
 import logging
 from pywren_ibm_cloud.version import __version__
 from pywren_ibm_cloud.config import cloud_logging_config
-from pywren_ibm_cloud.function import function_handler
-from pywren_ibm_cloud.function import function_invoker
+from pywren_ibm_cloud.worker import function_handler
+from pywren_ibm_cloud.worker import function_invoker
 
 cloud_logging_config(logging.INFO)
 logger = logging.getLogger('__main__')
