+++ pywren_ibm_cloud/worker/invoker.py
