+++ pywren_ibm_cloud/worker/utils.py
